pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "safestic"
version = "1.0.0"
description = "Safestic - Restic Backup Management Tool"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "colorama>=0.4.6",
    "requests>=2.31.0",
    "psutil>=5.9.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["services*"]
exclude = ["tests*", "examples*", "scripts*", "logs*", "temp*", "mount*", ".venv*", "__pycache__*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--cov=services --cov-report=term-missing"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "W", "I", "N", "D", "UP", "ANN", "B", "A"]
ignore = ["ANN101", "D203", "D213"]

[tool.ruff.pydocstyle]
convention = "numpy"

[tool.coverage.run]
source = ["services"]
omit = ["tests/*"]